+++ QAMGenFS2019/QAMGenFS2019.cproj
@@ -28,15 +28,15 @@
     <eraseonlaunchrule>0</eraseonlaunchrule>
     <AsfFrameworkConfig>
       <framework-data>
-  <options />
-  <configurations />
-  <files />
-  <documentation help="" />
-  <offline-documentation help="" />
-  <dependencies>
-    <content-extension eid="atmel.asf" uuidref="Atmel.ASF" version="3.34.1" />
-  </dependencies>
-</framework-data>
+        <options />
+        <configurations />
+        <files />
+        <documentation help="" />
+        <offline-documentation help="" />
+        <dependencies>
+          <content-extension eid="atmel.asf" uuidref="Atmel.ASF" version="3.34.1" />
+        </dependencies>
+      </framework-data>
     </AsfFrameworkConfig>
     <avrtool>com.atmel.avrdbg.tool.edbg</avrtool>
     <avrtoolserialnumber>ATMLEDBGC04201100001</avrtoolserialnumber>
@@ -282,6 +282,9 @@
     <Compile Include="includes\NHD0420Driver.h">
       <SubType>compile</SubType>
     </Compile>
+    <Compile Include="includes\protokollhandler.h">
+      <SubType>compile</SubType>
+    </Compile>
     <Compile Include="includes\utils.h">
       <SubType>compile</SubType>
     </Compile>
@@ -295,6 +298,9 @@
       <SubType>compile</SubType>
     </Compile>
     <Compile Include="NHD0420Driver.c">
+      <SubType>compile</SubType>
+    </Compile>
+    <Compile Include="protokollhandler.c">
       <SubType>compile</SubType>
     </Compile>
     <Compile Include="utils.c">
